moves SearchBar into layouts & adds temp testing files
