+++ package.json
@@ -23,13 +23,12 @@
     "react-dropzone-uploader": "^2.11.0",
     "react-multi-date-picker": "^3.3.0",
     "react-router-dom": "^6.3.0",
-    "react-scripts": "5.0.1",
-    "react-test-renderer": "^18.2.0"
+    "react-scripts": "5.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test",
+    "test": "react-scripts test --verbose --coverage",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
@@ -52,6 +51,7 @@
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^14.2.1",
+    "react-test-renderer": "^18.2.0",
     "sass": "^1.53.0"
   }
 }
