+++ src/layouts/DesktopNavBar.js
@@ -2,7 +2,7 @@ import React from "react";
 import { Link } from "react-router-dom";
 import { Typography } from "@mui/material";
 
-import { SearchBar } from "../components/SearchBar";
+import { SearchBar } from "../layouts/SearchBar";
 import { DesktopMenu } from "./DesktopMenu";
 export const DesktopNavBar = () => {
   return (
