+++ src/layouts/SearchBar.js
