--- src/test/AlertError.test.js
@@ -1,1 +1,0 @@
-// Test that error is received when making an invalid
