+++ src/test/SearchLocation.test.js
@@ -1,0 +1,12 @@
+// 1. User should receive at-least 1 charger when searching for Melbourne and Sydney
+// 2. User should receive 0 chargers when searching gibberish (eg. "eiorangioerangieoar") and a location not in Australia (eg. "Paris")
+// 3. Should receive the correct coordinates for Melbourne when executing the Geocoding function (-37.8136, 144.9631)
+
+import React from "react";
+import renderer from "react-test-renderer";
+
+describe("User search", () => {
+  it("Returns at-least 1 charger when searching 'Melbourne' or 'Sydney'", () => {
+    
+  });
+});
